Digitally sign the 'ghostscript printer' file ghostpdf.inf
Recent versions (8+) of Windows make it very difficult to install a device
whose driver is not signed. While the ghostpdf.inf file isn't really part
of Ghostscript we have had enquiries about producing s signed version. Now
that we have our own certificate, this is possible and is done here.

This should now install without problems on WIndows 8 and with fewer warnings
from UAC on Vista and Windows 7.