graphics library and interpreter changes to allow us to subclass the pattern accumulator device.
The main change is to pass a structure to the
device_spec_op which handle patterns. This structure carries the graphcs
state, which we require. This has required altering pdfwrite (the only
device currently implementing these spec_ops) and all the interpreters in
order to pass the required information.

An additional change places the GC structure descriptor for the pattern
accumulator device in the device structure's 'stype' instead of leaving it
NULL. We can't subclass devices with a NULL stype.