Consolidate almost all error codes into the graphics library
Do away with the separate/duplicate error code definitions in the Postscript
interpreter and in the graphics library. Also, where possible, do the same
for PCL.

The exception is the PXL code which has errors codes entirely of its own -
as unpleasant as that is, it would require rather more invasive changes to
have PXL use the common error codes. Possibly something to consider for a
future commit.

No cluster differences.