PDF Interpreter - fix a problem with DecodeParms
Bug #696019 "- Regression: reading PDF file produces error starting with ddaea8157085f4efb227e78326dd739d59493b97"

There was a problem whe we had an image dictionary with a /Filter, which was
not an array. We consumed the image and stream dictionaries which led to an
error.

This commit correctly updates and preserves the dictionaries.

No differences expected.