pdfwrite - fix consecutive CIE colour usage
Bug #695847 " colors in CIE color model are not converted correctly to PDF"

CIE colours cannot be represented directly in a PDF file, they *must* be
converted into an ICCBased colour space. If the colour conversion strategy
was LeaveColorUnchanged then the first usage of a colour in a CIEBased
colour space was correctly convrted, and the colour components rescaled,
but if the job then used another, different, colour in the same colour
space then the components were written out unscaled.

This commit forces colours in a CIEBased space to always be rescaled.

This gives one small progression in the test suite (09-31.ps) and some
slight colour differences in a further 7 files.