pdfwrite - write encrypted strings in Dests tree if protected PDF output
Bug #695992

We were forced to stop using names when creatng a Dests tree (instead of
an array), even though it seems legal, because Acrobat didn't like it.

However, I forgot to take account of password protected PDF files, which
require us to encrypt strings. Encryption is applied in this commit if
it is required.

No differences expected.