PDF interpreter - ignore broken SMask in ExtGState
Bug #696027 "Blank page instead of expected output"

The PDF file is invalid, it has an ExtGState with a SMask dictionary, the
SMask does not contain a /Group, which is required and the dictionary is
unusable without it.

This commit wraps the execution of SMask in a stopped so that if it fails
we can raise a warning and continue with the ExtGState, effectively ignoring
the invalid SMask.

This produces a small progression in tests_private/pdf/sumatra/broken_xobject_messing_up_clip_stack.pdf
that particular file is so damaged even Acrobat won't open it, but a
progression is nice anyway.