pdfwrite - fix ToUnicode emission
Bug #696021 "commit 24c562(fix for #692119) causes G2U mapping is lost in CIDFont"

Despite the bug title, it was in fact commit
1be6f4e4084443772b928510fb7ebce71572ea10 which broke the handling of the
code created by PSCRIPT5.DLL. The windows printing system emits a
GlyphNames2Unicode dictionary which you would expect to contain names and
Unicode values, but in fact contains character codes and Unicode values.

Sometimes we have the Character code to CID mapping being an identity, and
sometimes we don't. So we need to check both of these, which we can conveniently
do by lying to the decode_glyph procedure and pretending we are not a CIDFont.
We don't generally want to do this though, so we only do it if the CID and
character code differ. I haven't found a case where a PDF input file has a
differing CID and character code *and* fails decode_glyph with a CID, but
I suspect this may be possible, in which case we'll get a garbage ToUnicode
CMap, but it won't be worse than it was before

I think the code originally supplied by the user to fix this bug is dangerous
(may produce incorrect results with PDF input files) and so have not entirely
adopted it but created our own resolution.

No differences expected.