PDF interpreter - further work to handle illegal operations in text blocks
Bug #696017 "Portions of page missing"

The specimen file contains numerous operations illegally inside text blocks,
previously we fixed this for nested text blocks and more recently for images
inside text blocks. However this file also has stroke operations inside text
blocks (also illegal).

Here we handle this by redefining the path construction operators so that
they check for the existence of a SavedTextMatrix and use that instead of
the CTM if it is present. This has some consequences; we can't do this for
pattern PaintProcs or the pattern will be incorrectly defined, and the flag
we were previously using (presence of TextSavedMatrix in currentdict) can't
be used as that isn't present during the execution of a PaintProc. So we
have to look at the current definition of the marking operations (we choose
to just use /m for brevity) and take action based upon that. We always set
the 'normal' marking ops for a PaintProc, and if we are in a text block
we switch the ops back again after running the PaintProc.

No differences (except the obvious expected one in Bug696017.pdf)