Bug 696037: resolve issue with named pipes on Windows
The stat() call does not work for named pipes on Windows.

Move around the order of tests in lib_file_open_search_with_no_combine() so
it works with names pipes.

The way lib_file_open_search_with_no_combine() works is that we have a list of
paths to search, and list of paths from which we are allowed to open files.

If a file does not exist in one path from our search list, we continue
searching.

If, on the other hand, the file does exist, but is not in a path we are
permitted to use files from, we have to throw an invalidaccess error.

Previously, we'd stat() the file, then consult our allowed paths, then open the
file - taking appropriate alternative action if any of those stages failed.

Now, we consult the permitted paths first, and if that fails, only then do we
call stat() to decide whether to throw an error, or continue searching.

This should work correctly with Windows named pipes, as the names are always of
the form "\\\\.\\pipe\\..." (local) or "\\ServerName\pipe\..." (remote) so in
both cases the leading character being "\" will identify it as an absolute path
and those are always allowed.

No cluster differences.