PDF Interpreter - Don't attempt to draw annotations scaled to 0 size
Bug #696051 "Warning only appears with -dTextAlphaBits=4"

This commit does not address the underlying fault. The interpreter/graphics
library behaves differently with TextAlphaBits set and not. This:

/Helvetica findfont 20 scalefont setfont 10 10 moveto (test) show

gives an error with -dTextAlphaBits set, but not when it isn't. The latter
appears to be incorrect.

We work around this in the PDF interpreter by checking the scale factor
(because the actual fault is a /Rect [0 0 0 0]) and if either the x or y
scale is 0, we simply don't draw the annotation.

No differences expected.