saved-pages param processing was assuming the top device was the clist device
The parameter processing for saved-pages assumed that when the parameter
processing takes place the current device in the gstate is the rendering
device. If it has been subclassed this isn't the case (and what about
forwarding devices ?).

Have the code descend the chain of devices until reaching the terminal device
and use that, not the one in the gstate.