PDF interpreter - alter rebuilding of xref when problems are detected
Bug #696084 "Page rendered blank"

The file is pretty classically broken. It starts with an entry in the xref
for object 290 which is not flagged as 'free'. However, when we check the
file at the given offset it actually points to object 228.

This triggers us to rescan the file, because the xref is broken. So we sweep
the file from beginning to end looking for the sequence 'x y obj'. NB we
don't (can't) do this for object streams, because those are compressed.

When we find an object definition we enter it into the rebuilt xref table.

However, this file contains duplicate definitions for all the page objects
which leads us to emit a warning that there are multiple definitions and the
output may be incorrect. The rebuilding code uses the *last* definition of
a given object and the highest generation number, because in general when
there are multiple definitions of the same object, the last one is the one
that's correct (PDF tools usually modify the file by appending rather than
writing into the middle of the file).

However, in this case, the *first* definition is the correct one. Not only
that but the second definitions aren't even valid.....

So the problem is caused by; a broken xref table, multiple definitions of
the same object, and the later definitions being invalid.

I've worked around this by checking the *original* xref table and if one of
the definitions is at the same offset as defined in the original xref we now
always prefer that definition, otherwise we prefer the last definition, as
before.

No differences expected.