PDF interpreter - check type of XObject resoruces when scanning for transparency
Bug #696083 "Error reading PDF file"

The PDF file has a single page, but the /Pages dictionary has a Resources
dictionary (which is consulted if any page is missing a resource definition)

The Page is complete, but when scanning for the presence of transparency
(to improve performance) we test any 'Parent' resources as well, and in this
case the Pages dictionary contains an XObject resource which is incorrect.

Instead of a dictoinary, its an integer, and since that's the wrong type we
throw an error.

This commit checks the type of XObject resources during scanning for transparency
and ignores (with a warning) any which have an incorrect type.

No differences expected.