Update TrueType font copying code to handle OpenType 1.7 OS/2 tables
Bug #696049 "Postscript can not konverted to PDF"

The embedded TrueType font conforms to version 1.7 of the OpenType specifiction
which introduces 2 new variables to the OS/2 table. This makes the OS/2 table
length too great for the sanity checks in the code and it abrots with an error.

This commit extends the checking to cater for version 1.7 OpenType OS/2
tables, and removes an unused variable at the same time.

No differences expected.