pdfwrite - More fiddling with symbolic TrueType fonts
Bug #695969 "Regression: Missing umlauts starting with 28b7d3c75adb8e942f2158bdec445a147f945530"

The problem here is the way that Acrobat behaves with symbolic fonts. If we define
a font as 'symbolic', then we don't write an Encoding (its not permitted
with PDF/A, and is technically invalid), and Acrobat uses the font's 1,0
CMAP subtable.

If we do not write the font as symbolic then we need to write an Encoding. In
this case it seems that Acrobat uses the embedded font's 3,1 CMAP subtable.

The problem is that the two tables need not be consistent, and in the case
of this file, are not. The character code 77 maps to GID 48 in both cases
but the character code 252 maps to GID 129 (a composite glyph) with the 3,1
CMAP and to a /.notdef with the 1,0 CMAP.

This is only a problem if we *don't* subset the font, because if we subset the
font we *generate* a set of CMAP subtables, and we generate tables which
are consistent with each other.

I did consider forcing the CMAP tables to be generated, even when not subsetting
the font, but this font contains more than 255 glyphs, so we can't fit them
all into a 1,0 CMAP anyway, so which ones should we use ?

So instead, if we have a TrueType font, are embedding it, are *not* subsetting it
*and* the original font was 'not symbolic', (which we test by looking for the
presence of a 'nearest_encoding_index' in the original font as this is set
from the is_symbolic flag in the PDF interpreter, which is set from the
original font descriptor Flags value), then we do *NOT* mark the font as
symbolic.

This also means altering the check against creating an Encoding, if we are
(eventually) going to make the font non-symbolic, then we need to add an
Encoding to it as well.

This causes a difference in one cluster test, this is due to the fact that
the font can't be embedded due to licencing restrictions. Because of this
the output was already incorrect, its now incorrect in a slightly different
way.