Revise VS sln and vcproj to avoid parallel builds.
Parallel builds from the VS GUI don't work because the multiple nmake instances
are unaware of each others' activities (hence builds will fail when more than
one process attempts to write to a file).

To address this, add an "all" project which builds all the executables, and mark
the individual projects as "don't build".

Thus each executable can be built indvidually by selecting the project in the
solution explorer, and explicitly selecting build for that project. OTOH,
building the whole solution only builds the "all" project, which automatically
builds each executable in turn.