MS solution; add 'All' project.
The All project invokes some new makefile targets. These makefile
targets just depend on all the appropriate existing makefile targets.
Thus we get a sequential build of the different projects.

Nmake does not (yet) support parallelism, so we don't have the problem
of building multiple projects in parallel as we get when we build
multiple projects in the solution.

We use the configuration manager in the solution to ensure that when
we build the entire solution, only 'All' is built.