PDF interpreter - clamp horizontal text spacing to prevent degenerate CTM
Bug #696107 "Text missing with -dTextAlphaBits=4"

The file contains three occurrences of "0 Tz" which sets the horizontal
scale factor to 0. This causes the CTM to have a zero horizontal scale
which (correctly) throws an error when TextAlphaBits is set.

This commit detects setting Tz to 0 and replaces it with a miniscule, but
non-zero, value thus preventing the degenerate matrix and the error.

A later commit will change the behaviour of the graphics library so that it
is at least consistent with differing values for TextAlphaBits.