Bug 696101: fix uses of the sfopen API.
The stream API in GS is defined as *always* opening files in binary mode,
where applicable, so there is no need for the API clients to specify binary
mode.

This is previously been benign, and thus ignored, but reportedly ending up with
a duplicate 'b' character in the mode causes a crash on Windows 10.

No cluster differences.