Initial commit for color management in gproof device.
There are four possible cases that had to be handled.
They depend upon if we have or do not have non standard colorants
and if we have or do not have the proper ICC profiles

Essentially we may have to compute the equivalent CMYK colors
for the image that includes spot colors and we may have to do
a very slow conversion to RGB when we don't have the proper
ICC profiles.   The fastest case, occurs when we have an ICC
for all our source colorants (including any spots).