Force the normal text handling to throw errors under the same (degenerate CTM) conditions as when -dTextAlphaBits is set.
Currently if the CTM is something like [0 0 0 0 Tx Ty] then Ghostscript
throws an error if -dTextAlphaBits is set, but not if it isn't. This leads
to a number of bug reports citing problems with TextAlphaBits which really
isn't the case.

Testing Adobe seems to indicate that, for PostScript, and error is the
expected behaviour.

This commit adds a 'currentpoint' call in two places so that we get the same
error no matter which code path we follow.

The PDF interpreter has already been fixed so that all known cases will not
throw errors with this commit, but two PCL files from Quality Logic do now
throw an error, 18-08.bin and 18-09.bin. A bug report will be raised for
these.