Ensure imager state is not marked as graphics state
In various places in the pdf14 compositor device, we use local copies of the
imager state (so stuff specific to the pdf14 compositor functions doesn't
accidentally affect the "upstream" state).

This ensures that those copies are not marked as being a graphics state, in case
the imager state started life as a full graphics state.

No cluster differences.