Fix for bug 696124, character scaling with a singular matrix.
By experiment HP does not ignore requests for font height which are out
of range, instead it clamps the values to the limits of the range.  The
minimum is .25 points.  The cet files in these bugs are requesting a
font size of 0 and the HP laserjet does render a very small character
which is visible but only readable with magnification assuming normal
vision.

This change just allows the matrix to be scaled to the minimum value
allowing the test files in the bug to continue processing, these tiny
characters do not render properly even with this change.  The rounding of
escapement would have to be worked in PCL and likely the font rasterizer
would need to be looked at, the characters do not appear correct even
with proper spacing.

There is no practical demand to do this correctly, so we don't plan to
work on it anytime soon.