PDF interpreter - Attempt to validate the ICCBased /N value for color spaces
Bug #696120 "Invalid PDF to TIFF conversion"

The supplied specimen file is incorrect, it has an ICCBased colour space
which contains the /N key (number of channels) set to a value of 3, but
the actual profile is a gray profile, and therefore /N should be 1.

Here we add a new function .numicc_components which attempts to find the
number of components from an ICC profile (returns 0 if it can't determine
the value). We then use that returned value to compare to the actual
/N value given. If the two don't match, and the calculated value is non-zero
then we use the calculated value instead, and alter the value of /N.

One file in the cluster test suite is shown as different, but it is visually
the same.

In passing, remove some unused variables and commented out code from zicc.c
to silence some compiler warnings.