Recursive make calls use -f "makefile name" option
The recursive make calls previously relied
on the makefile being the default name "Makefile" - by passing the name of
the makefile to the sub-call, we can use custom makefile names.

Currently only works for GNU make (detected in configure).

For this to work, if you rename the makefile, you must also edit the name in
the "MAKEFILE" variable to match your custom makefile name.

No cluster differences.