pdfwrite - consider both X and Y axes when deciding on downsampling
Bug #696152 "gs no more rendering figure inside the frame"

The decision on downsampling is taken by comparing the image resolution
against the downsampling threshold. Previously we only considered the
resolution in the X direction, assuming (not entirely unreasonably) that
the image would be of similar resolution in each direction.

However in this case the image is massively scaled (and therefore has a
practical very high resolution) in the X direction, but is barely scaled
at all in the Y direction.

Downsampling this image caused the height of the image to be reduced to 1
whcih effectively caused it to disappear.

No differences expected.