vector devices - prevent the use of Graphics and TextAlphaBits
The GraphicsAlphaBits and TextAlphaBits parameters don't make sense and don't
work with the vector devices, with the exception of rendering transparency.
In addition they can potentially cause seg faults and other errors.

Although we've always had code to detect the use of this at rendering
time, this hasn't stopped people from pointlessly using them.

To prevent such foolishness (at the expense of transparency flattening, which
is a very minor usage) this commit prints a warning and throws an error on
attempts to use these switches with vector devices.

Additionally update the documentation to make it clear that this is not
permitted.