Bug 696161: ensure we have a show enumerator object
before using it as such - it is possible (via an invalid configuration) to get
a text enumerator, and treating that as a show enumerator can cause a crash.

Similarly, ensure we have a graphics state (and not just an imager state).

Found and fix suggested by Ken.

No cluster differences.