pdfwrite - track soft mask IDs in our saved graphcs states
Bug #695250 "Missing text from pdfwrite"

The problem was that we were not tracking soft mask IDs in our saved copies
of the graphics state, which meant that we were not properly adding (or
removing) when creating ExtGStates in pdf_update_alpha().

This addresses the specific issue in the bug, as well as showing progressions
in the test file x_-_renders_slowly.pdf (missing Qt logo now appears)

However, a more ambitious treatment, which better tracks gsave/grestore by
actually emitting q/Q operations seems to show still further improvement.
I'm going to add this as a separate commit as, despite the progressions,
I'm a little unsure about it.