Fixes PCL crash when a new ICC directory is specified on the command line.
The library context must use the same allocator to create and free the
icc directory string.  The memory parameter passed in is not necessarily
the same as the original.  For example the PCL client wraps the
allocator to use the chunk allocator after the icc directory string is
created (part of gs_malloc_init).  This can result in memory being
managed by 2 different memory handlers.  Fortunately the original memory
pointer was tucked in the library context structure and that can be used
instead of the parameter passed in.

There are other areas in the library context code that need to be
checked for this same problem, and a discussion about how this might be
handled better is in the works.