Bug 696178: avoid var overflow with large CFF font
If a CFF/Type 2 font has a large number of (g)subrs, the remaining buffer size
when we come to recreate the font for Freetype could be larger than 64435
bytes, and this could overflow the usigned short parameter we pass to the
helper function which extracts and stores each (g)subr.

As I doubt we'll see a font with a single (g)subr of more than 65535 bytes, I'm
just clamping the value to the usigned short range.

No cluster differences.