PJL Parser - change a number of static tables to dynamic allocations
In order to provide for storage in excess of 256 bytes for a given PJL
variable, we either need to increase the size of the fixed strings stored
in a number of tales, or change the string storage to be dynamic.

Obviously its more memory efficient (at the cost of some performance) to
allocate the strings to the required size rather than wasting memory.

This commit alters the PJL 'environment' as well as its defaults and
associated factory defaults table, and the similar fontsource and its
defaults and associated factory defaults tables so that we only allocate
memory as required.