PJL interpreter - add new methods to configure pdfwrite
Bug #693117 "PCL -> PDF/A"
Bug #693058 "There is no way to control many pdfwrite features in GhostPCL (eg AutoRotatePages)"

Originally developed on the branch PJL_pdfwrite_config, full history
is available on that branch.

Added new PDFMARK and SETDISTILLERPARAMS PJL tokens which will allow
for 'PostScript like' syntax whch can be used to configure the
pdfwrite family of devices.

This will allow for the production of PDF/A-1b files and also the
control of some features which cannot be set from the command line.

Updated the documentation, gathering all the high level devices in
one place, split the documentation into controls based on the
input language and the output format.

No differences expected.