Add 'boiler plate' for PCL->PDF/A conversion
This adds a new PJL option 'PDFA_PROFILE' which when set should contain a
UTF-8 path and filename for an RGB ICC profile.

The remaining code adds all the boiler plate required to actually create
all the various objects required for a PDF/A OutputIntent.

To use this the user must also set -dPDFA=1 on the command line. Currently
we assume that the output will be RGB so its not valid to set a
ColorConversionStrategy other than LeaveColorUnchanged or RGB.