Fix memory leak with pattern clists.
The clist reader's ICC cache gets recreated every time we execute a clist,
but previously the code did not take action to clean it up, this the
reference count was wrong, and we leaked memory.

Whilst it may be preferable to have it persist, in the short term, it seems
safer to cleanup after clist playback, as that is functionally the same as the
existing behaviour.

No cluster differences