Fix pdf14 compositor/ICC reference counting problems.
One reference counting issue whereby the code originally assumed it would be
moving an existing ICC profile reference, rather than creating a new reference
(thus no reference counting was done) and it transpired that in certain cases
a new profile could end up being created - thus reference counting was required
to ensure that, in the case where a newly created profile was used, it would
also be freed as required. In this case, one of the places where the reference
was removed did not decrement the reference count.

Secondly, in the event of the compositor being interrupted (by an error or user
intervention) between push group and pop group operations, when we come to
finalize the compositor device, we have to emulate the ICC profile restorations
that happen during a pop group, so all the references and reference counts end
up back where they were when the push group happened.

No cluster differences.