Prevent crash with reference count debugging (-Z^)
The garbage collector makes no promises about the order in which objects are
swept. As a result, it is possible for a forwarding device to remain unswept
after its target has already.

To handle such relationships, the garbage collector leaves memory contents alone
until the entire sweep is complete, so the forwarding device's decrement of the
target reference count is safe. However, with rc debugging active, the debugging
code attempts to ascertain the object type by interrogating the memory header,
which the garbage has, by now, userped for it's own uses - causing a crash.

By setting the rc.memory pointer to NULL when the target device is finalized,
the rc debugging code is kept happy.

In the non-gc case, the finalize is the very last thing to happen before the
memory is actually freed, thus the memory pointer is moot by that stage.

No cluster differences.