eps2write - convert path bounding box to user space from device space
Bug #696181 "BoundingBox wrong when using eps2write in AIX environment"

I can't test this on an AIX setup, but for me the test file was incorrect
on Windows anyway.

In order to find the marking extent of a (filled) path we intersect the
path with the current clipping path and check the extents of the resulting
rectangle against the extents of the current BBox.

However, the result of the intersection is in device space, and we did not
'undo' the resolution. Other places in the code *do* remove the scaling
due to resolution, so we simply adopt that here.

No differences expected.