Bug 694180 (related): jbig2dec base streams error handling
Two problems:

First: when creating the jbig2dec context, you optionally pass it a function
pointer and data pointer to use to record error conditions. The data pointer
being used was the pointer to the Ghostscript stream state - which was allocated
in normal Postscript memory - i.e. subject to relocation by the garbage
collector. As this was stored as part of jbig2dec's internals, there was no
way for the garbage collector to move the pointer, and thus the location
was potentially nonsense.

Secondly, in testing the fix for above, it turns out that, in the event of an
error during JBIG2 decoding, there was no cleanup done - only the memory for
the stream state would be discarded (by the garbage collector), the jbig2dec
context was left to leak.

To fix the first problem, add a non-gc memory structure with the pointers
required for the jbig2dec error callback.

For the second problem, add a finalize method to the jbig2dec stream state
so the non-gc stuff can be cleaned up in the event of an error.

No cluster differences.