pdfwrite - yet more symbolic font meddling
Bug #696207 "Creation of PDF/A-1b from PDF results in missing characters"

The previous fix for symbolic fonts could lead to a situation where we emitted
an embedded, subset, symbolic TrueType font, bu put a /Encoding in the font.

This is invalid for PDF/A-1 and also causes Acrobat to do strange things with
the CMAP subtable(s) in the font. This commit should ensure (by synchronising
the tests) that we don't emit symbolic TT fonts with an Encoding.

Checking with the test file for bug 695969 indicates that the code still
works for that case too.

This has unfortunatley required some re-organisation of data structure
declarations.

I think we won't be able to reliably have this work until the TrueType font
embedding code is rewritten.

In any event the heuristics seem improved again.