Process pdfmark operations inserted via PJL into putdeviceparams
This allows something similar to the general pdfmark syntax to be used in
PJL in order to control various aspects of PDF creation. Examples include
creating the PDF/A-1 objects required to conform to that standard, and the
addition of document information such as the Creator.

The syntax is not precisely the same as for PostScript;

Each token must be preceded by a space, so "<< /Key /Value >>" is legal
while "<</Key/Value>>" is not even though it would be in PostScript.

PostScript operators cannot be used, so "(test) cvn" will not work, if you
want a name you need to define it as such. Therefore names with spaces
cannot be defined.

Its no longer possible to create a file object and /PUT it on a predefined
stream. This is vital for PDF/A-1 creation, so we add a new /PUTFILE action.
This takes a predefined stream object, and a string containing the name of
the file whose content is to be written to that stream object.

Other syntax changes may be required, this is an initial commit, and its likely
that more work will need to be done on this.