Bug 696223: cope with empty %rom% returning an error
The romfs is always included now, but if there are no files (i.e. if we're using
the "dummy" romfs for COMPILE_INITS=0) it returns an "unregistered" error from
the "status" method.

This trips up the code for identifying the ICC profiles path, which stats the
default ICC profile path, which happens to be %rom%iccprofiles.

So, put the status call in a stopped context, and act accordingly.

No cluster differences.