Avoid doing color linearity tests when they are not needed.
In the shading code there are several tests to determine if we
are sufficiently linear in our fill region.  If not, it is further
divided.  For certain color mappings we are always going to be linear
and do not need to check.  We test for these now and avoid the check.
Cluster push reported 119 diffs.  Checked them and they looked
reasonable.  Differences in various gradient fills.  The run time
numbers from the regression test improved significantly.
Going from a cpu time of 37:29:22  to 29:28:04 with gs going from
19:42:34 to 15:25:29