Bug 696229: initialize clist color space validity.
The clist uses a stack variable to hold the device color. Valgrind reports
that the ccolor_valid entry in the device color can remain unset, so set
it to invalid when the initialize the device color to "unset".

Also, add a PACIFY_VALGRIND memset so pattern bitmap buffers are fully zeroed
on allocation - stops valgrind getting upset about writing and reading back
uninitialised (padding) bytes to/from the clist.

No cluster differences.