pdfwrite - ensure we don't write non-zero matrix elements with too tiny a value
Bug #696228 "Regression: error starting with 70b442162ff8f3d40595d0eb9fb365d341139ee2"

The change noted causes the PDF interpreter to replace degenerate text matrices
with a very tiny scaling matrix. However, when writing this back through
ps2write, we also take the size of the font into account, and bake this
into the matrix.

If the font size was less than 1 this could lead to us writing a value too
small for printf's '%g' format, and it ended up being written as zero, which
caused the error with the ps2write output.

This commit checks each of the array elements to see if they are non-zero, but
have a value so small that they will *become* zero when written out. If this
happens we recalculate the scale so that the minimum value is emitted. The
value is so small anyway that this should not cause a perceptible difference.
We *don't* do this for the Tx and Ty values as we want these to be correct
and it doesn't cause a problem if these are 0.

No differences expected.