pdfwrite - prevent NULL pointer dereference
Bug #696234 " Regression: segfault with pdfwrite starting with 17131c7a7fdf9d8537c4715e538c49b29c8945a8"

When trying to find an object in the resource chains using an ID, we can
encounter resources which have been cancelled. These have no object (it has
been freed) so trying to dereference the object can cause a SEG fault.

I did try to free the resource as well, when cancelling it but that spun
off into a maze of complications involving the garbage collector. It seems
that 'something' was maintaining a pointer to those structures, so freeing
them caused all kinds of relocation/free problems.

Its a lot easier just to prevent dereferencing the pointer.

No differences expected.