PS interpreter - fix a buffer overrun in '=='
Since changing the implementation of sprintf we now have a situation
where %g can return > 50 bytes of data for a INF. In obj_cvp we define a
fixed size buffer on the heap to receive this, but only 50 bytes wide, so
the buffer overflows and we get a crash.

Increased the size of the buffer well past the maximum possible return
size from sprintf.

No differences expected.