PS interpreter - check for floating point exceptions in mul, div and add
Bug #696248 "gs9.18 crash on display of nan; also poor handling of inf"

The mul, div and add operators did not detect invalid results (infinity or
non-a-number). This is likely because there is no guaranteed portable
method for detection of these prior to C99.

Here we add checks for this results using the C99 extensions isnan() and
isinf(). These are guarded by #ifdef so that they do not cause compilation
problems on compilers that don't support them. Additionally add support for
early versions of Microsoft Visual Studio which used somewhat different
names for these functions.

This code requires some changes to the configure scripts on Linus in
order to enable it.

On systems which do not support isnan() and isinf() there will be no change,
we will continue to not report errors in ths case.

No differences expected.