prevent rinkj device crash when misconfigured (no SetupFile)
Bug #696246 "Ghostscript 9.18 with -dFirstPage/-dLastPage fails for ijs and some x11 devices"

The rinkj device requires a SetupFile to be given as a device parameter,
however it doesn't actually check to see if one is given, and just tries
to open the filename, with a predictable crash when none is given.

Here we check the filename and attempt to ensure it is both present and
minimally valid.

No differences expected.