Commit of WOFF font support for GS
Support in Ghostscript is implemented by augmenting the TrueType handling in
the Postscript world. Loading WOFF fonts "stripped" is not supported, except
for the purpose of finding the font name from the name table.

There is also a C implementation in the graphics library which takes a memory
buffer or a stream, and unpacks the WOFF into a TTF in a memory buffer. It
is currently not called.