pdfwrite - guard against NULL pointer dereference and correct a loop
Bug #696275 "-dFastWebView parameter crashes Ghostscript"

This is a twofold fix. A Coverity static analysis fix added some return
checking, but this happened inside a faulty loop, the code should not have
been executed. Because the return value wasn't checked, this didn't cause a
problem......

Checking the return value meant that we returned an error to the caller, and
an oversight there could lead to us dereferencing a pointer which had not
been allocated.

This commit fixes both of these problems.

No differences expected.