PDF interpreter - ignore indirect references to invalid object number 0
Bug #696289 "**** Unrecoverable error in ref! writing PDF file"

The Outlines tree is referenced indirectly as "0 0 R" which is invalid
(object number 0 is reserved). The resolveR routine which resolves indirect
references checks the validity of the object number but was checking for a
value *less* than 0, when it should have been less than or equal.

This commit alters the lt to le, which allows the file to be processed, of
course there are no Outlines present since those were originally broken.

Note this only exhibits with high level devices because rendering devices
cannot usefully use the Outlines tree, and ignore it.


No differences expected.